move validate config to config utils
