+++ packages/jest-puppeteer-env/src/configUtils.js
@@ -65,7 +65,60 @@ const getPuppeteer = (config) => {
   }
 };
 
+const validateConfig = (config) => {
+  if (!config.rootDir) {
+    throw new Error('You need to specify the `rootDir` in your jest config!');
+  } else if (!config.testEnvironmentOptions) {
+    throw new Error('You need to specify the `testEnvironmentOptions` in your jest config!');
+  } else if (!config.testEnvironmentOptions.mockResponseDir) {
+    throw new Error(
+      'Please specify where the mocks should be saved to and loaded from using the `mockResponseDir` test environment option.'
+    );
+  } else if (config.testEnvironmentOptions.collectCoverage && !config.testEnvironmentOptions.coverageDirectory) {
+    throw new Error('When coverage is collected you need to provide a coverageDirectory option.');
+  } else if (config.testEnvironmentOptions.recordScreenshots && !config.testEnvironmentOptions.screenshotDirectory) {
+    throw new Error('When screenshots are taken you need to provide a screenshotDirectory option.');
+  } else if (config.testEnvironmentOptions.collectPerfMetrics && !config.testEnvironmentOptions.perfMetricsDirectory) {
+    throw new Error('When performance metrics are collected you need to provide a perfMetricsDirectory option.');
+  } else {
+    const {
+      rootDir,
+      testEnvironmentOptions: {
+        collectCoverage,
+        collectCoverageFrom,
+        coverageDirectory,
+        mockResponseDir,
+        perfMetricsDirectory,
+        printCoverageSummary,
+        recordCoverageText,
+        recordScreenshots,
+        requestPathIgnorePatterns,
+        screenshotDirectory,
+        shouldUseMocks,
+        collectPerfMetrics
+      }
+    } = config;
+
+    return {
+      collectCoverage,
+      collectCoverageFrom,
+      collectPerfMetrics,
+      coverageDirectory,
+      mockResponseDir,
+      perfMetricsDirectory,
+      printCoverageSummary,
+      recordCoverageText,
+      recordScreenshots,
+      requestPathIgnorePatterns,
+      rootDir,
+      screenshotDirectory,
+      shouldUseMocks
+    };
+  }
+};
+
 module.exports = {
   getPuppeteer,
-  readConfig
+  readConfig,
+  validateConfig
 };
